SAMZA-2068: Separating container launch logic into util class
